requirements.txt
bump2version
git+https://github.com/polyswarm/malwarerepoclient.git#egg=malwarerepoclient
polyswarm-artifact
pytest==3.9.2
pytest-asyncio==0.9.0
yara-python
